requirements.txt
fastapi>=0.95.0
uvicorn[standard]>=0.22.0
SQLAlchemy>=2.0.0
typer>=0.9.0
python-dotenv>=1.0.0
requests>=2.0.0
beautifulsoup4>=4.0.0
jinja2>=3.0.0
python-multipart>=0.0.5